Split stage.py monolith into focused per-stage files (SemanticDedup pattern)
extraction.py:   DripperHTMLExtractionStage + MinerU bindings (~373 lines)
inference.py:    DripperHTMLInferenceStage + shared inference helpers (~312 lines)
preprocessing.py: DripperHTMLPreprocess + DripperHTMLPostprocessStage (~473 lines)
stage.py:        shared utils + DripperHTMLLayoutTemplateStage (~2811 lines)

Key refactor: promoted 7 static methods from DripperHTMLExtractionStage
(_coerce_html, _coerce_optional_str, _sanitize_case_output_html,
_get_processed_attr, _case_has_item_ids, _count_item_ids,
_is_empty_document_error) to module-level functions in stage.py so all
sub-modules can share them without circular imports.

All sub-modules import shared utilities from stage.py; __init__.py
re-exports all public classes so external import paths are unchanged.

Matches SemanticDedup structure: kmeans.py, pairwise.py, workflow.py focused.
All files pass py_compile and ruff check.
