+++ nemo_curator/stages/text/experimental/dripper/__init__.py
@@ -17,15 +17,22 @@
 Requirements:
     pip install "nemo-curator[dripper]"
     # Installs: mineru-html>=1.1, llm-web-kit>=4.1
+
+Module layout:
+    stage.py          — shared utilities + DripperHTMLLayoutTemplateStage
+    extraction.py     — DripperHTMLExtractionStage + MinerU bindings
+    inference.py      — DripperHTMLInferenceStage
+    preprocessing.py  — DripperHTMLPreprocessStage + DripperHTMLPostprocessStage
+    workflow.py       — DripperHTMLWorkflow (high-level entry point)
 """
 
-from nemo_curator.stages.text.experimental.dripper.stage import (
-    DripperHTMLExtractionStage,
-    DripperHTMLInferenceStage,
-    DripperHTMLLayoutTemplateStage,
+from nemo_curator.stages.text.experimental.dripper.extraction import DripperHTMLExtractionStage
+from nemo_curator.stages.text.experimental.dripper.inference import DripperHTMLInferenceStage
+from nemo_curator.stages.text.experimental.dripper.preprocessing import (
     DripperHTMLPostprocessStage,
     DripperHTMLPreprocessStage,
 )
+from nemo_curator.stages.text.experimental.dripper.stage import DripperHTMLLayoutTemplateStage
 from nemo_curator.stages.text.experimental.dripper.workflow import DripperHTMLWorkflow
 
 __all__ = [
