+++ nemo_curator/stages/text/experimental/dripper/stage.py
@@ -12,7 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Dripper HTML main-content extraction through Curator inference clients."""
+"""Dripper HTML main-content extraction through Curator inference clients.
+
+Shared utilities and DripperHTMLLayoutTemplateStage live here.
+Stage classes are split across focused sub-modules:
+  extraction.py   — DripperHTMLExtractionStage
+  inference.py    — DripperHTMLInferenceStage
+  preprocessing.py — DripperHTMLPreprocessStage, DripperHTMLPostprocessStage
+"""
 
 from __future__ import annotations
 
@@ -397,1059 +404,88 @@ def _rebuild_batch(batch: DocumentBatch, df: pd.DataFrame) -> DocumentBatch:
     return new_batch
 
 
-@dataclass(kw_only=True)
-class DripperHTMLExtractionStage(ProcessingStage[DocumentBatch, DocumentBatch]):
-    """Extract main HTML/content with Dripper through a Curator LLM client."""
-
-    name: str = "DripperHTMLExtractionStage"
-    client: AsyncLLMClient | None
-    model_name: str
-    html_col: str = "html"
-    url_col: str | None = "url"
-    output_html_col: str = "dripper_html"
-    output_content_col: str = "dripper_content"
-    raw_response_col: str = "dripper_response"
-    preprocess_time_col: str = "dripper_preprocess_time_s"
-    inference_time_col: str = "dripper_inference_time_s"
-    postprocess_time_col: str = "dripper_postprocess_time_s"
-    total_time_col: str = "dripper_time_s"
-    error_col: str = "dripper_error"
-    warning_col: str = "dripper_warning"
-    item_count_col: str = "dripper_item_count"
-    prompt_chars_col: str = "dripper_prompt_chars"
-    request_max_tokens_col: str = "dripper_request_max_tokens"
-    prompt_tokens_col: str = "dripper_prompt_tokens"
-    completion_tokens_col: str = "dripper_completion_tokens"
-    total_tokens_col: str = "dripper_total_tokens"
-    prompt_version: str = "short_compact"
-    output_format: str = "mm_md"
-    fallback: Literal["trafilatura", "bypass", "empty"] = "trafilatura"
-    generation_config: GenerationConfig | None = None
-    dynamic_max_tokens: bool = False
-    dynamic_max_token_padding: int = 16
-    dynamic_max_tokens_per_item: int = 6
-    dynamic_min_max_tokens: int = 32
-    structured_output_mode: Literal["none", "structured_outputs", "guided_regex"] = "none"
-    max_concurrent_requests: int = 64
-    health_check: bool = True
-    keep_intermediate: bool = False
-    simplified_html_col: str = "dripper_simplified_html"
-    mapped_html_col: str = "dripper_mapped_html"
-
-    _bindings: _MinerUHTMLBindings | None = field(init=False, repr=False, default=None)
-    _fallback_handler: Any = field(init=False, repr=False, default=None)
-    _initialized: bool = field(init=False, repr=False, default=False)
-
-    def __post_init__(self) -> None:
-        if self.client is None:
-            msg = "DripperHTMLExtractionStage requires a non-None 'client' (AsyncLLMClient)"
-            raise ValueError(msg)
-        self.model_name = self.model_name.strip()
-        if not self.model_name:
-            msg = "DripperHTMLExtractionStage requires a non-empty 'model_name'"
-            raise ValueError(msg)
-        if self.max_concurrent_requests <= 0:
-            msg = "max_concurrent_requests must be positive"
-            raise ValueError(msg)
-        if self.dynamic_max_token_padding < 0:
-            msg = "dynamic_max_token_padding must be non-negative"
-            raise ValueError(msg)
-        if self.dynamic_max_tokens_per_item <= 0:
-            msg = "dynamic_max_tokens_per_item must be positive"
-            raise ValueError(msg)
-        if self.dynamic_min_max_tokens <= 0:
-            msg = "dynamic_min_max_tokens must be positive"
-            raise ValueError(msg)
-        if self.structured_output_mode not in _STRUCTURED_OUTPUT_MODES:
-            msg = f"structured_output_mode must be one of {sorted(_STRUCTURED_OUTPUT_MODES)}"
-            raise ValueError(msg)
-
-    def inputs(self) -> tuple[list[str], list[str]]:
-        return ["data"], [self.html_col]
-
-    def outputs(self) -> tuple[list[str], list[str]]:
-        columns = [
-            self.output_html_col,
-            self.output_content_col,
-            self.raw_response_col,
-            self.preprocess_time_col,
-            self.inference_time_col,
-            self.postprocess_time_col,
-            self.total_time_col,
-            self.error_col,
-            self.warning_col,
-            self.item_count_col,
-            self.prompt_chars_col,
-            self.request_max_tokens_col,
-            self.prompt_tokens_col,
-            self.completion_tokens_col,
-            self.total_tokens_col,
-        ]
-        if self.keep_intermediate:
-            columns.extend([self.simplified_html_col, self.mapped_html_col])
-        return ["data"], columns
-
-    def setup(self, worker_metadata: WorkerMetadata | None = None) -> None:  # noqa: ARG002
-        if self._initialized:
-            return
-
-        self._bindings = _load_mineru_html_bindings()
-        self._fallback_handler = self._bindings.get_fallback_handler(self.fallback)
-        self.client.setup()
-        if self.health_check:
-            self._run_health_check()
-        self._initialized = True
-
-    def process(self, batch: DocumentBatch) -> DocumentBatch:
-        if not self._initialized:
-            self.setup()
-
-        df = batch.to_pandas().copy()
-        if self.html_col not in df.columns:
-            msg = f"Input batch is missing required HTML column: {self.html_col!r}"
-            raise ValueError(msg)
-
-        html_values = df[self.html_col].tolist()
-        if self.url_col is not None and self.url_col in df.columns:
-            url_values = df[self.url_col].tolist()
-        else:
-            url_values = [None] * len(df)
-
-        results = run_async_safe(lambda: self._extract_all_async(html_values, url_values))
-        df[self.output_html_col] = [r.main_html for r in results]
-        df[self.output_content_col] = [r.main_content for r in results]
-        df[self.raw_response_col] = [r.raw_response for r in results]
-        df[self.preprocess_time_col] = [r.preprocess_time_s for r in results]
-        df[self.inference_time_col] = [r.inference_time_s for r in results]
-        df[self.postprocess_time_col] = [r.postprocess_time_s for r in results]
-        df[self.total_time_col] = [r.total_time_s for r in results]
-        df[self.error_col] = [r.error for r in results]
-        df[self.warning_col] = [r.warning for r in results]
-        df[self.item_count_col] = [r.item_count for r in results]
-        df[self.prompt_chars_col] = [r.prompt_chars for r in results]
-        df[self.request_max_tokens_col] = [r.request_max_tokens for r in results]
-        df[self.prompt_tokens_col] = [r.prompt_tokens for r in results]
-        df[self.completion_tokens_col] = [r.completion_tokens for r in results]
-        df[self.total_tokens_col] = [r.total_tokens for r in results]
-        if self.keep_intermediate:
-            df[self.simplified_html_col] = [r.simplified_html for r in results]
-            df[self.mapped_html_col] = [r.mapped_html for r in results]
-
-        return _rebuild_batch(batch, df)
-
-    def _run_health_check(self) -> None:
-        run_async_safe(lambda: _run_dripper_health_check(self.client, self.model_name, self.generation_config))
-
-    async def _extract_all_async(self, html_values: list[object], url_values: list[object]) -> list[_DripperRowResult]:
-        sem = asyncio.Semaphore(self.max_concurrent_requests)
-
-        async def _extract_one_throttled(html_value: object, url_value: object) -> _DripperRowResult:
-            async with sem:
-                return await self._extract_one_async(html_value, url_value)
-
-        tasks = [
-            _extract_one_throttled(html_value, url_value)
-            for html_value, url_value in zip(html_values, url_values, strict=False)
-        ]
-        raw_results = await asyncio.gather(*tasks, return_exceptions=True)
-
-        results: list[_DripperRowResult] = []
-        for idx, result in enumerate(raw_results):
-            if isinstance(result, BaseException):
-                logger.error("Dripper extraction failed for row {}: {}", idx, result)
-                results.append(_DripperRowResult(error=str(result)))
-            else:
-                results.append(result)
-        return results
-
-    def _preprocess_case(self, case: object) -> tuple[object, int, str, str, bool]:
-        """Simplify HTML, count items, build prompt. Returns (case, item_count, prompt, warning, needs_llm)."""
-        case = self._bindings.simplify_single_input(case)
-        item_count = self._count_item_ids(case)
-        if not self._case_has_item_ids(case):
-            case = self._bindings.extract_main_html_fallback(case, fallback_handler=self._fallback_handler)
-            return (
-                case,
-                item_count,
-                "",
-                "no _item_id attributes after simplification; used fallback without LLM",
-                False,
-            )
-        case = self._bindings.build_prompt(case, prompt_version=self.prompt_version)
-        prompt = case.generate_input.full_prompt
-        return case, item_count, prompt, "", True
-
-    async def _run_inference_async(
-        self, case: object, prompt: str, item_count: int
-    ) -> tuple[object, str, int, int, int, int]:
-        """Run inference and postprocess. Returns (case, raw_response, request_max_tokens, prompt_tokens, completion_tokens, total_tokens)."""
-        generation_config = _with_structured_output_config(
-            self._generation_config_for_item_count(item_count), prompt, self.structured_output_mode
-        )
-        request_max_tokens = generation_config.max_tokens or 0
-        raw_response, prompt_tokens, completion_tokens, total_tokens = await _query_dripper_model(
-            self.client, self.model_name, [{"role": "user", "content": prompt}], generation_config
-        )
-        case.generate_output = self._bindings.generate_output_cls(response=raw_response)
-        case = self._bindings.parse_result(case)
-        case = self._bindings.extract_main_html_single(case)
-        return case, raw_response, request_max_tokens, prompt_tokens, completion_tokens, total_tokens
-
-    async def _extract_one_async(self, html_value: object, url_value: object) -> _DripperRowResult:
-        start_total = time.perf_counter()
-        html = self._coerce_html(html_value)
-        if not html.strip():
-            return _DripperRowResult(total_time_s=time.perf_counter() - start_total, warning="empty HTML input")
-
-        url = self._coerce_optional_str(url_value)
-        case = self._bindings.case_cls(self._bindings.input_cls(raw_html=html, url=url))
-        raw_response = ""
-        preprocess_time_s = 0.0
-        inference_time_s = 0.0
-        postprocess_time_s = 0.0
-        primary_error = ""
-        warning = ""
-        item_count = 0
-        prompt_chars = 0
-        request_max_tokens = 0
-        prompt_tokens = 0
-        completion_tokens = 0
-        total_tokens = 0
-
-        try:
-            start_preprocess = time.perf_counter()
-            case, item_count, prompt, warning, needs_llm = self._preprocess_case(case)
-            preprocess_time_s = time.perf_counter() - start_preprocess
-            if needs_llm:
-                prompt_chars = len(prompt)
-                start_inference = time.perf_counter()
-                (
-                    case,
-                    raw_response,
-                    request_max_tokens,
-                    prompt_tokens,
-                    completion_tokens,
-                    total_tokens,
-                ) = await self._run_inference_async(case, prompt, item_count)
-                inference_time_s = time.perf_counter() - start_inference
-                start_postprocess = time.perf_counter()
-                postprocess_time_s += time.perf_counter() - start_postprocess
-        except Exception as exc:  # noqa: BLE001
-            if preprocess_time_s == 0.0:
-                preprocess_time_s = time.perf_counter() - start_total
-            primary_error = str(exc)
-            logger.debug("Dripper primary extraction failed, applying {} fallback: {}", self.fallback, primary_error)
-            try:
-                start_fallback = time.perf_counter()
-                case = self._bindings.extract_main_html_fallback(case, fallback_handler=self._fallback_handler)
-                postprocess_time_s += time.perf_counter() - start_fallback
-                warning = primary_error
-            except Exception as fallback_exc:  # noqa: BLE001
-                error = f"{primary_error}; fallback failed: {fallback_exc}"
-                return _DripperRowResult(
-                    raw_response=raw_response,
-                    preprocess_time_s=preprocess_time_s,
-                    inference_time_s=inference_time_s,
-                    postprocess_time_s=postprocess_time_s,
-                    total_time_s=time.perf_counter() - start_total,
-                    error=error,
-                    warning=primary_error,
-                    simplified_html=self._get_processed_attr(case, "simpled_html"),
-                    mapped_html=self._get_processed_attr(case, "map_html"),
-                    item_count=item_count,
-                    prompt_chars=prompt_chars,
-                    request_max_tokens=request_max_tokens,
-                    prompt_tokens=prompt_tokens,
-                    completion_tokens=completion_tokens,
-                    total_tokens=total_tokens,
-                )
-
-        conversion_error, postprocess_time_s = self._convert_extraction_output(case, postprocess_time_s)
-        base = _DripperRowResult(
-            raw_response=raw_response,
-            preprocess_time_s=preprocess_time_s,
-            inference_time_s=inference_time_s,
-            postprocess_time_s=postprocess_time_s,
-            total_time_s=time.perf_counter() - start_total,
-            warning=warning,
-            simplified_html=self._get_processed_attr(case, "simpled_html"),
-            mapped_html=self._get_processed_attr(case, "map_html"),
-            item_count=item_count,
-            prompt_chars=prompt_chars,
-            request_max_tokens=request_max_tokens,
-            prompt_tokens=prompt_tokens,
-            completion_tokens=completion_tokens,
-            total_tokens=total_tokens,
-        )
-        return self._build_extraction_result(case, base, conversion_error=conversion_error)
-
-    def _convert_extraction_output(self, case: object, postprocess_time_s: float) -> tuple[str, float]:
-        conversion_error = ""
-        start_conversion = time.perf_counter()
-        try:
-            self._sanitize_case_output_html(case)
-            case = self._bindings.convert2content(case, output_format=self.output_format)
-            postprocess_time_s += time.perf_counter() - start_conversion
-        except Exception as exc:  # noqa: BLE001
-            postprocess_time_s += time.perf_counter() - start_conversion
-            conversion_error = str(exc)
-            logger.debug("Dripper content conversion failed: {}", conversion_error)
-        return conversion_error, postprocess_time_s
-
-    def _build_extraction_result(
-        self, case: object, base: _DripperRowResult, *, conversion_error: str
-    ) -> _DripperRowResult:
-        output_data = getattr(case, "output_data", None)
-        main_html = getattr(output_data, "main_html", "") if output_data is not None else ""
-        main_content = getattr(output_data, "main_content", "") if output_data is not None else ""
-        if main_content is None:
-            main_content = ""
-        error = ""
-        warning = base.warning
-        if conversion_error:
-            if self._is_empty_document_error(conversion_error) and not str(main_html).strip():
-                warning = _append_warning(warning, conversion_error)
-            else:
-                error = conversion_error
-        return replace(base, main_html=main_html, main_content=main_content, error=error, warning=warning)
-
-    @staticmethod
-    def _sanitize_case_output_html(case: object) -> None:
-        output_data = getattr(case, "output_data", None)
-        if output_data is None:
-            return
-        main_html = getattr(output_data, "main_html", None)
-        if isinstance(main_html, str):
-            output_data.main_html = _strip_xml_incompatible_chars(main_html)
-
-    @staticmethod
-    def _get_processed_attr(case: object, attr: str) -> str:
-        process_data = getattr(case, "process_data", None)
-        value = getattr(process_data, attr, "") if process_data is not None else ""
-        return value if isinstance(value, str) else ""
-
-    @classmethod
-    def _case_has_item_ids(cls, case: object) -> bool:
-        return "_item_id" in cls._get_processed_attr(case, "simpled_html") or "_item_id" in cls._get_processed_attr(
-            case,
-            "map_html",
-        )
-
-    @classmethod
-    def _count_item_ids(cls, case: object) -> int:
-        html = cls._get_processed_attr(case, "simpled_html") or cls._get_processed_attr(case, "map_html")
-        return len(set(_ITEM_ID_RE.findall(html)))
-
-    def _generation_config_for_item_count(self, item_count: int) -> GenerationConfig:
-        base = self.generation_config or GenerationConfig()
-        if not self.dynamic_max_tokens or base.max_tokens is None or item_count <= 0:
-            return base
-
-        dynamic_max_tokens = max(
-            self.dynamic_min_max_tokens,
-            item_count * self.dynamic_max_tokens_per_item + self.dynamic_max_token_padding,
-        )
-        return replace(base, max_tokens=min(base.max_tokens, dynamic_max_tokens))
-
-    @staticmethod
-    def _coerce_html(value: object) -> str:
-        if _is_missing(value):
-            return ""
-        if isinstance(value, bytes | bytearray):
-            raw_bytes = bytes(value)
-            decoded = _decode_html_bytes(raw_bytes)
-            if decoded is None:
-                decoded = raw_bytes.decode("utf-8", errors="replace")
-            return _strip_xml_incompatible_chars(decoded or "")
-        return _strip_xml_incompatible_chars(str(value))
-
-    @staticmethod
-    def _coerce_optional_str(value: object) -> str | None:
-        if _is_missing(value):
-            return None
-        text = str(value)
-        return text if text else None
-
-    @staticmethod
-    def _is_empty_document_error(error: str) -> bool:
-        normalized = error.lower()
-        return "document is empty" in normalized or "empty html tree" in normalized or "empty html input" in normalized
-
-
-class DripperHTMLPreprocessStage(ProcessingStage[DocumentBatch, DocumentBatch]):
-    """Simplify HTML and build Dripper prompts before model inference."""
-
-    name: str = "DripperHTMLPreprocessStage"
-    html_col: str = "html"
-    url_col: str | None = "url"
-    raw_response_col: str = "dripper_response"
-    preprocess_time_col: str = "dripper_preprocess_time_s"
-    inference_time_col: str = "dripper_inference_time_s"
-    postprocess_time_col: str = "dripper_postprocess_time_s"
-    total_time_col: str = "dripper_time_s"
-    error_col: str = "dripper_error"
-    warning_col: str = "dripper_warning"
-    item_count_col: str = "dripper_item_count"
-    prompt_chars_col: str = "dripper_prompt_chars"
-    request_max_tokens_col: str = "dripper_request_max_tokens"
-    prompt_tokens_col: str = "dripper_prompt_tokens"
-    completion_tokens_col: str = "dripper_completion_tokens"
-    total_tokens_col: str = "dripper_total_tokens"
-    simplified_html_col: str = "dripper_simplified_html"
-    mapped_html_col: str = "dripper_mapped_html"
-    prompt_version: str = "short_compact"
-    generation_config: GenerationConfig | None = None
-    dynamic_max_tokens: bool = False
-    dynamic_max_token_padding: int = 16
-    dynamic_max_tokens_per_item: int = 6
-    dynamic_min_max_tokens: int = 32
-    worker_count: int | None = None
-
-    _bindings: _MinerUHTMLBindings | None = field(init=False, repr=False, default=None)
-    _initialized: bool = field(init=False, repr=False, default=False)
-
-    def __post_init__(self) -> None:
-        if self.dynamic_max_token_padding < 0:
-            msg = "dynamic_max_token_padding must be non-negative"
-            raise ValueError(msg)
-        if self.dynamic_max_tokens_per_item <= 0:
-            msg = "dynamic_max_tokens_per_item must be positive"
-            raise ValueError(msg)
-        if self.dynamic_min_max_tokens <= 0:
-            msg = "dynamic_min_max_tokens must be positive"
-            raise ValueError(msg)
-        if self.worker_count is not None and self.worker_count <= 0:
-            msg = "worker_count must be positive when set"
-            raise ValueError(msg)
-
-    def num_workers(self) -> int | None:
-        return self.worker_count
-
-    def inputs(self) -> tuple[list[str], list[str]]:
-        return ["data"], [self.html_col]
-
-    def outputs(self) -> tuple[list[str], list[str]]:
-        return ["data"], [
-            self.raw_response_col,
-            self.preprocess_time_col,
-            self.inference_time_col,
-            self.postprocess_time_col,
-            self.total_time_col,
-            self.error_col,
-            self.warning_col,
-            self.item_count_col,
-            self.prompt_chars_col,
-            self.request_max_tokens_col,
-            self.prompt_tokens_col,
-            self.completion_tokens_col,
-            self.total_tokens_col,
-            self.simplified_html_col,
-            self.mapped_html_col,
-            _DRIPPER_PROMPT_COL,
-            _DRIPPER_NEEDS_LLM_COL,
-            _DRIPPER_PRIMARY_ERROR_COL,
-            _DRIPPER_EMPTY_INPUT_COL,
-        ]
-
-    def setup(self, worker_metadata: WorkerMetadata | None = None) -> None:  # noqa: ARG002
-        if self._initialized:
-            return
-        self._bindings = _load_mineru_html_bindings()
-        self._initialized = True
-
-    def process(self, batch: DocumentBatch) -> DocumentBatch:
-        if not self._initialized:
-            self.setup()
-
-        df = batch.to_pandas().copy()
-        if self.html_col not in df.columns:
-            msg = f"Input batch is missing required HTML column: {self.html_col!r}"
-            raise ValueError(msg)
-
-        html_values = df[self.html_col].tolist()
-        if self.url_col is not None and self.url_col in df.columns:
-            url_values = df[self.url_col].tolist()
-        else:
-            url_values = [None] * len(df)
-
-        results = [
-            self._prepare_one(html_value, url_value)
-            for html_value, url_value in zip(html_values, url_values, strict=False)
-        ]
-
-        df[self.raw_response_col] = ""
-        df[self.preprocess_time_col] = [r.preprocess_time_s for r in results]
-        df[self.inference_time_col] = 0.0
-        df[self.postprocess_time_col] = 0.0
-        df[self.total_time_col] = [r.preprocess_time_s for r in results]
-        df[self.error_col] = ""
-        df[self.warning_col] = [r.warning for r in results]
-        df[self.item_count_col] = [r.item_count for r in results]
-        df[self.prompt_chars_col] = [r.prompt_chars for r in results]
-        df[self.request_max_tokens_col] = [r.request_max_tokens for r in results]
-        df[self.prompt_tokens_col] = 0
-        df[self.completion_tokens_col] = 0
-        df[self.total_tokens_col] = 0
-        df[self.simplified_html_col] = [r.simplified_html for r in results]
-        df[self.mapped_html_col] = [r.mapped_html for r in results]
-        df[_DRIPPER_PROMPT_COL] = [r.prompt for r in results]
-        df[_DRIPPER_NEEDS_LLM_COL] = [r.needs_llm for r in results]
-        df[_DRIPPER_PRIMARY_ERROR_COL] = [r.primary_error for r in results]
-        df[_DRIPPER_EMPTY_INPUT_COL] = [r.empty_input for r in results]
-
-        self._log_metrics(
-            {
-                "preprocess_rows": float(len(df)),
-                "preprocess_llm_rows": float(sum(r.needs_llm for r in results)),
-                "preprocess_fallback_rows": float(sum((not r.needs_llm) and (not r.empty_input) for r in results)),
-            }
-        )
-        return _rebuild_batch(batch, df)
-
-    def _prepare_one(self, html_value: object, url_value: object) -> _DripperPrepResult:
-        started = time.perf_counter()
-        html = DripperHTMLExtractionStage._coerce_html(html_value)
-        if not html.strip():
-            return _DripperPrepResult(
-                empty_input=True,
-                preprocess_time_s=time.perf_counter() - started,
-                warning="empty HTML input",
-            )
-
-        url = DripperHTMLExtractionStage._coerce_optional_str(url_value)
-        case = self._bindings.case_cls(self._bindings.input_cls(raw_html=html, url=url))
-        simplified_html = ""
-        mapped_html = ""
-        item_count = 0
-        try:
-            case = self._bindings.simplify_single_input(case)
-            simplified_html = DripperHTMLExtractionStage._get_processed_attr(case, "simpled_html")
-            mapped_html = DripperHTMLExtractionStage._get_processed_attr(case, "map_html")
-            item_count = DripperHTMLExtractionStage._count_item_ids(case)
-            if not DripperHTMLExtractionStage._case_has_item_ids(case):
-                return _DripperPrepResult(
-                    needs_llm=False,
-                    preprocess_time_s=time.perf_counter() - started,
-                    warning="no _item_id attributes after simplification; used fallback without LLM",
-                    simplified_html=simplified_html,
-                    mapped_html=mapped_html,
-                    item_count=item_count,
-                )
-
-            case = self._bindings.build_prompt(case, prompt_version=self.prompt_version)
-            prompt = case.generate_input.full_prompt
-            generation_config = self._generation_config_for_item_count(item_count)
-            return _DripperPrepResult(
-                prompt=prompt,
-                needs_llm=True,
-                preprocess_time_s=time.perf_counter() - started,
-                simplified_html=simplified_html,
-                mapped_html=mapped_html,
-                item_count=item_count,
-                prompt_chars=len(prompt),
-                request_max_tokens=generation_config.max_tokens or 0,
-            )
-        except Exception as exc:  # noqa: BLE001
-            primary_error = str(exc)
-            logger.debug("Dripper preprocessing failed; postprocess stage will apply fallback: {}", primary_error)
-            return _DripperPrepResult(
-                needs_llm=False,
-                preprocess_time_s=time.perf_counter() - started,
-                primary_error=primary_error,
-                warning=primary_error,
-                simplified_html=simplified_html,
-                mapped_html=mapped_html,
-                item_count=item_count,
-            )
-
-    def _generation_config_for_item_count(self, item_count: int) -> GenerationConfig:
-        return DripperHTMLExtractionStage._generation_config_for_item_count(self, item_count)
-
-
-@dataclass(kw_only=True)
-class DripperHTMLInferenceStage(ProcessingStage[DocumentBatch, DocumentBatch]):
-    """Run only Dripper model inference against an OpenAI-compatible client."""
-
-    name: str = "DripperHTMLInferenceStage"
-    client: AsyncLLMClient | None
-    model_name: str
-    raw_response_col: str = "dripper_response"
-    inference_time_col: str = "dripper_inference_time_s"
-    warning_col: str = "dripper_warning"
-    item_count_col: str = "dripper_item_count"
-    request_max_tokens_col: str = "dripper_request_max_tokens"
-    prompt_tokens_col: str = "dripper_prompt_tokens"
-    completion_tokens_col: str = "dripper_completion_tokens"
-    total_tokens_col: str = "dripper_total_tokens"
-    generation_config: GenerationConfig | None = None
-    structured_output_mode: Literal["none", "structured_outputs", "guided_regex"] = "none"
-    max_concurrent_requests: int = 64
-    health_check: bool = False
-    worker_count: int | None = None
-
-    _initialized: bool = field(init=False, repr=False, default=False)
-
-    def __post_init__(self) -> None:
-        if self.client is None:
-            msg = "DripperHTMLInferenceStage requires a non-None 'client' (AsyncLLMClient)"
-            raise ValueError(msg)
-        self.model_name = self.model_name.strip()
-        if not self.model_name:
-            msg = "DripperHTMLInferenceStage requires a non-empty 'model_name'"
-            raise ValueError(msg)
-        if self.max_concurrent_requests <= 0:
-            msg = "max_concurrent_requests must be positive"
-            raise ValueError(msg)
-        if self.structured_output_mode not in _STRUCTURED_OUTPUT_MODES:
-            msg = f"structured_output_mode must be one of {sorted(_STRUCTURED_OUTPUT_MODES)}"
-            raise ValueError(msg)
-        if self.worker_count is not None and self.worker_count <= 0:
-            msg = "worker_count must be positive when set"
-            raise ValueError(msg)
-
-    def num_workers(self) -> int | None:
-        return self.worker_count
-
-    def inputs(self) -> tuple[list[str], list[str]]:
-        return ["data"], [_DRIPPER_PROMPT_COL, _DRIPPER_NEEDS_LLM_COL, self.request_max_tokens_col]
-
-    def outputs(self) -> tuple[list[str], list[str]]:
-        return ["data"], [
-            self.raw_response_col,
-            self.inference_time_col,
-            self.warning_col,
-            self.prompt_tokens_col,
-            self.completion_tokens_col,
-            self.total_tokens_col,
-            _DRIPPER_PRIMARY_ERROR_COL,
-        ]
-
-    def setup(self, worker_metadata: WorkerMetadata | None = None) -> None:  # noqa: ARG002
-        if self._initialized:
-            return
-        self.client.setup()
-        if self.health_check:
-            run_async_safe(lambda: _run_dripper_health_check(self.client, self.model_name, self.generation_config))
-        self._initialized = True
-
-    def process(self, batch: DocumentBatch) -> DocumentBatch:
-        if not self._initialized:
-            self.setup()
-
-        df = batch.to_pandas().copy()
-        results = run_async_safe(lambda: self._infer_all_async(df))
-
-        needs_llm = df[_DRIPPER_NEEDS_LLM_COL].astype(bool).tolist()
-        existing_raw_responses = (
-            df[self.raw_response_col].astype(str).tolist() if self.raw_response_col in df else [""] * len(df)
-        )
-        existing_inference_times = (
-            pd.to_numeric(df[self.inference_time_col], errors="coerce").fillna(0.0).tolist()
-            if self.inference_time_col in df
-            else [0.0] * len(df)
-        )
-        existing_prompt_tokens = (
-            pd.to_numeric(df[self.prompt_tokens_col], errors="coerce").fillna(0).astype(int).tolist()
-            if self.prompt_tokens_col in df
-            else [0] * len(df)
-        )
-        existing_completion_tokens = (
-            pd.to_numeric(df[self.completion_tokens_col], errors="coerce").fillna(0).astype(int).tolist()
-            if self.completion_tokens_col in df
-            else [0] * len(df)
-        )
-        existing_total_tokens = (
-            pd.to_numeric(df[self.total_tokens_col], errors="coerce").fillna(0).astype(int).tolist()
-            if self.total_tokens_col in df
-            else [0] * len(df)
-        )
-        existing_warnings = df[self.warning_col].astype(str) if self.warning_col in df else pd.Series([""] * len(df))
-        existing_primary_errors = (
-            df[_DRIPPER_PRIMARY_ERROR_COL].astype(str)
-            if _DRIPPER_PRIMARY_ERROR_COL in df
-            else pd.Series([""] * len(df))
-        )
-        df[self.raw_response_col] = [
-            r.raw_response if should_query else existing_raw
-            for r, should_query, existing_raw in zip(results, needs_llm, existing_raw_responses, strict=True)
-        ]
-        df[self.inference_time_col] = [
-            r.inference_time_s if should_query else existing_time
-            for r, should_query, existing_time in zip(results, needs_llm, existing_inference_times, strict=True)
-        ]
-        df[self.warning_col] = [
-            _append_warning(existing_warning, result.warning)
-            for existing_warning, result in zip(existing_warnings.tolist(), results, strict=True)
-        ]
-        df[_DRIPPER_PRIMARY_ERROR_COL] = [
-            _append_warning(existing_error, result.primary_error)
-            for existing_error, result in zip(existing_primary_errors.tolist(), results, strict=True)
-        ]
-        df[self.prompt_tokens_col] = [
-            r.prompt_tokens if should_query else existing_tokens
-            for r, should_query, existing_tokens in zip(results, needs_llm, existing_prompt_tokens, strict=True)
-        ]
-        df[self.completion_tokens_col] = [
-            r.completion_tokens if should_query else existing_tokens
-            for r, should_query, existing_tokens in zip(results, needs_llm, existing_completion_tokens, strict=True)
-        ]
-        df[self.total_tokens_col] = [
-            r.total_tokens if should_query else existing_tokens
-            for r, should_query, existing_tokens in zip(results, needs_llm, existing_total_tokens, strict=True)
-        ]
-
-        llm_prompts = [
-            str(row.get(_DRIPPER_PROMPT_COL, "") or "")
-            for _, row in df.iterrows()
-            if bool(row.get(_DRIPPER_NEEDS_LLM_COL, False))
-        ]
-        non_empty_llm_prompts = [prompt for prompt in llm_prompts if prompt.strip()]
-        unique_llm_prompts = len(set(non_empty_llm_prompts))
-        self._log_metrics(
-            {
-                "inference_rows": float(len(df)),
-                "inference_llm_rows": float(sum(bool(v) for v in df[_DRIPPER_NEEDS_LLM_COL].tolist())),
-                "inference_unique_llm_prompts": float(unique_llm_prompts),
-                "inference_dedup_saved_rows": float(len(non_empty_llm_prompts) - unique_llm_prompts),
-                "inference_errors": float(sum(1 for r in results if r.primary_error)),
-            }
-        )
-        return _rebuild_batch(batch, df)
-
-    async def _infer_all_async(self, df: pd.DataFrame) -> list[_DripperInferenceResult]:
-        sem = asyncio.Semaphore(self.max_concurrent_requests)
-        prompts = df[_DRIPPER_PROMPT_COL].astype(str).tolist()
-        needs_llm = df[_DRIPPER_NEEDS_LLM_COL].astype(bool).tolist()
-        request_max_tokens = (
-            pd.to_numeric(df[self.request_max_tokens_col], errors="coerce").fillna(0).astype(int).tolist()
-            if self.request_max_tokens_col in df.columns
-            else [0] * len(df)
-        )
-
-        async def _infer_one_throttled(prompt: str, row_max_tokens: int) -> _DripperInferenceResult:
-            async with sem:
-                return await self._infer_one_async(prompt, True, row_max_tokens)
-
-        grouped_indexes: dict[tuple[str, int], list[int]] = defaultdict(list)
-        results: list[_DripperInferenceResult | None] = [None] * len(df)
-        for idx, (prompt, should_query, row_max_tokens) in enumerate(
-            zip(prompts, needs_llm, request_max_tokens, strict=True)
-        ):
-            if not should_query:
-                results[idx] = _DripperInferenceResult()
-            elif not prompt.strip():
-                results[idx] = _DripperInferenceResult(
-                    primary_error="empty Dripper prompt", warning="empty Dripper prompt"
-                )
-            else:
-                grouped_indexes[(prompt, row_max_tokens)].append(idx)
-
-        tasks = {key: _infer_one_throttled(prompt=key[0], row_max_tokens=key[1]) for key in grouped_indexes}
-        raw_results = await asyncio.gather(*tasks.values(), return_exceptions=True)
-
-        for (_key, indexes), result in zip(grouped_indexes.items(), raw_results, strict=True):
-            if isinstance(result, BaseException):
-                logger.error("Dripper inference failed for prompt group {} rows: {}", len(indexes), result)
-                error = str(result)
-                first_result = _DripperInferenceResult(primary_error=error, warning=error)
-            else:
-                first_result = result
-            first_idx = indexes[0]
-            results[first_idx] = first_result
-            for duplicate_idx in indexes[1:]:
-                results[duplicate_idx] = replace(
-                    first_result,
-                    inference_time_s=0.0,
-                    prompt_tokens=0,
-                    completion_tokens=0,
-                    total_tokens=0,
-                )
-
-        return [result if result is not None else _DripperInferenceResult() for result in results]
-
-    async def _infer_one_async(self, prompt: str, should_query: bool, row_max_tokens: int) -> _DripperInferenceResult:
-        if not should_query:
-            return _DripperInferenceResult()
-        if not prompt.strip():
-            return _DripperInferenceResult(primary_error="empty Dripper prompt", warning="empty Dripper prompt")
-
-        started = time.perf_counter()
-        try:
-            generation_config = self.generation_config or GenerationConfig()
-            if row_max_tokens > 0 and generation_config.max_tokens != row_max_tokens:
-                generation_config = replace(generation_config, max_tokens=row_max_tokens)
-            generation_config = _with_structured_output_config(generation_config, prompt, self.structured_output_mode)
-            raw_response, prompt_tokens, completion_tokens, total_tokens = await self._query_model_with_usage(
-                model=self.model_name,
-                messages=[{"role": "user", "content": prompt}],
-                generation_config=generation_config,
-            )
-        except Exception as exc:  # noqa: BLE001
-            error = str(exc)
-            logger.debug("Dripper inference failed; postprocess stage will apply fallback: {}", error)
-            return _DripperInferenceResult(
-                inference_time_s=time.perf_counter() - started,
-                primary_error=error,
-                warning=error,
-            )
-        return _DripperInferenceResult(
-            raw_response=raw_response,
-            inference_time_s=time.perf_counter() - started,
-            prompt_tokens=prompt_tokens,
-            completion_tokens=completion_tokens,
-            total_tokens=total_tokens,
-        )
-
-    async def _query_model_with_usage(
-        self,
-        *,
-        model: str,
-        messages: list[dict[str, str]],
-        generation_config: GenerationConfig,
-    ) -> tuple[str, int, int, int]:
-        query_model_with_usage = getattr(self.client, "query_model_with_usage", None)
-        if callable(query_model_with_usage):
-            response = await query_model_with_usage(
-                model=model,
-                messages=messages,
-                generation_config=generation_config,
-            )
-            contents = getattr(response, "contents", [])
-            return (
-                contents[0] if contents else "",
-                _coerce_usage_int(getattr(response, "prompt_tokens", None)),
-                _coerce_usage_int(getattr(response, "completion_tokens", None)),
-                _coerce_usage_int(getattr(response, "total_tokens", None)),
-            )
-
-        response = await self.client.query_model(
-            model=model,
-            messages=messages,
-            generation_config=generation_config,
-        )
-        return response[0] if response else "", 0, 0, 0
-
-
-@dataclass(kw_only=True)
-class DripperHTMLPostprocessStage(ProcessingStage[DocumentBatch, DocumentBatch]):
-    """Parse Dripper responses, extract main HTML, and convert content."""
-
-    name: str = "DripperHTMLPostprocessStage"
-    html_col: str = "html"
-    url_col: str | None = "url"
-    output_html_col: str = "dripper_html"
-    output_content_col: str = "dripper_content"
-    raw_response_col: str = "dripper_response"
-    preprocess_time_col: str = "dripper_preprocess_time_s"
-    inference_time_col: str = "dripper_inference_time_s"
-    postprocess_time_col: str = "dripper_postprocess_time_s"
-    total_time_col: str = "dripper_time_s"
-    error_col: str = "dripper_error"
-    warning_col: str = "dripper_warning"
-    fallback: Literal["trafilatura", "bypass", "empty"] = "trafilatura"
-    output_format: str = "mm_md"
-    keep_intermediate: bool = False
-    simplified_html_col: str = "dripper_simplified_html"
-    mapped_html_col: str = "dripper_mapped_html"
-    worker_count: int | None = None
-
-    _bindings: _MinerUHTMLBindings | None = field(init=False, repr=False, default=None)
-    _fallback_handler: Any = field(init=False, repr=False, default=None)
-    _initialized: bool = field(init=False, repr=False, default=False)
-
-    def __post_init__(self) -> None:
-        if self.worker_count is not None and self.worker_count <= 0:
-            msg = "worker_count must be positive when set"
-            raise ValueError(msg)
-
-    def num_workers(self) -> int | None:
-        return self.worker_count
-
-    def inputs(self) -> tuple[list[str], list[str]]:
-        return ["data"], [
-            self.html_col,
-            self.raw_response_col,
-            self.simplified_html_col,
-            self.mapped_html_col,
-            _DRIPPER_NEEDS_LLM_COL,
-            _DRIPPER_PRIMARY_ERROR_COL,
-            _DRIPPER_EMPTY_INPUT_COL,
-        ]
-
-    def outputs(self) -> tuple[list[str], list[str]]:
-        columns = [
-            self.output_html_col,
-            self.output_content_col,
-            self.postprocess_time_col,
-            self.total_time_col,
-            self.error_col,
-            self.warning_col,
-        ]
-        if self.keep_intermediate:
-            columns.extend([self.simplified_html_col, self.mapped_html_col])
-        return ["data"], columns
-
-    def setup(self, worker_metadata: WorkerMetadata | None = None) -> None:  # noqa: ARG002
-        if self._initialized:
-            return
-        self._bindings = _load_mineru_html_bindings()
-        self._fallback_handler = self._bindings.get_fallback_handler(self.fallback)
-        self._initialized = True
-
-    def process(self, batch: DocumentBatch) -> DocumentBatch:
-        if not self._initialized:
-            self.setup()
-
-        df = batch.to_pandas().copy()
-        html_values = df[self.html_col].tolist()
-        if self.url_col is not None and self.url_col in df.columns:
-            url_values = df[self.url_col].tolist()
-        else:
-            url_values = [None] * len(df)
-
-        results = [
-            self._postprocess_one(row, html_value, url_value)
-            for (_, row), html_value, url_value in zip(df.iterrows(), html_values, url_values, strict=True)
-        ]
-
-        preprocess_times = _numeric_series_or_zero(df, self.preprocess_time_col)
-        inference_times = _numeric_series_or_zero(df, self.inference_time_col)
-        postprocess_times = pd.Series([r.postprocess_time_s for r in results], index=df.index)
-
-        df[self.output_html_col] = [r.main_html for r in results]
-        df[self.output_content_col] = [r.main_content for r in results]
-        df[self.postprocess_time_col] = postprocess_times
-        df[self.total_time_col] = preprocess_times + inference_times + postprocess_times
-        df[self.error_col] = [r.error for r in results]
-        df[self.warning_col] = [r.warning for r in results]
-
-        drop_cols = [
-            _DRIPPER_PROMPT_COL,
-            _DRIPPER_NEEDS_LLM_COL,
-            _DRIPPER_PRIMARY_ERROR_COL,
-            _DRIPPER_EMPTY_INPUT_COL,
-            _DRIPPER_LAYOUT_FINALIZED_COL,
-        ]
-        if not self.keep_intermediate:
-            drop_cols.extend([self.simplified_html_col, self.mapped_html_col])
-        df = df.drop(columns=[col for col in drop_cols if col in df.columns])
-
-        self._log_metrics(
-            {
-                "postprocess_rows": float(len(df)),
-                "postprocess_errors": float(sum(1 for r in results if r.error)),
-                "postprocess_warnings": float(sum(1 for r in results if r.warning)),
-            }
-        )
-        return _rebuild_batch(batch, df)
-
-    def _postprocess_one(self, row: pd.Series, html_value: object, url_value: object) -> _DripperPostResult:
-        started = time.perf_counter()
-        warning = str(row.get(self.warning_col, "") or "")
-        primary_error = str(row.get(_DRIPPER_PRIMARY_ERROR_COL, "") or "")
-        if bool(row.get(_DRIPPER_LAYOUT_FINALIZED_COL, False)):
-            return _DripperPostResult(
-                main_html=str(row.get(self.output_html_col, "") or ""),
-                main_content=row.get(self.output_content_col, "") or "",
-                postprocess_time_s=float(row.get(self.postprocess_time_col, 0.0) or 0.0),
-                error=str(row.get(self.error_col, "") or ""),
-                warning=warning,
-            )
-        html = DripperHTMLExtractionStage._coerce_html(html_value)
-        if bool(row.get(_DRIPPER_EMPTY_INPUT_COL, False)) or not html.strip():
-            return _DripperPostResult(
-                postprocess_time_s=time.perf_counter() - started,
-                warning=warning or "empty HTML input",
-            )
-
-        url = DripperHTMLExtractionStage._coerce_optional_str(url_value)
-        case = self._build_case(
-            html=html,
-            url=url,
-            simplified_html=str(row.get(self.simplified_html_col, "") or ""),
-            mapped_html=str(row.get(self.mapped_html_col, "") or ""),
-        )
-        raw_response = str(row.get(self.raw_response_col, "") or "")
-        needs_llm = bool(row.get(_DRIPPER_NEEDS_LLM_COL, False))
-
-        case, warning, fallback_error = self._postprocess_prepare_case(
-            case,
-            raw_response=raw_response,
-            needs_llm=needs_llm,
-            primary_error=primary_error,
-            warning=warning,
-        )
-        if fallback_error:
-            return _DripperPostResult(
-                postprocess_time_s=time.perf_counter() - started,
-                error=fallback_error,
-                warning=warning,
-            )
-
-        conversion_error = ""
-        try:
-            DripperHTMLExtractionStage._sanitize_case_output_html(case)
-            case = self._bindings.convert2content(case, output_format=self.output_format)
-        except Exception as exc:  # noqa: BLE001
-            conversion_error = str(exc)
-            logger.debug("Dripper content conversion failed: {}", conversion_error)
-
-        output_data = getattr(case, "output_data", None)
-        main_html = getattr(output_data, "main_html", "") if output_data is not None else ""
-        main_content = getattr(output_data, "main_content", "") if output_data is not None else ""
-        if main_content is None:
-            main_content = ""
-        error = ""
-        if conversion_error:
-            if DripperHTMLExtractionStage._is_empty_document_error(conversion_error) and not str(main_html).strip():
-                warning = _append_warning(warning, conversion_error)
-            else:
-                error = conversion_error
-
-        return _DripperPostResult(
-            main_html=main_html,
-            main_content=main_content,
-            postprocess_time_s=time.perf_counter() - started,
-            error=error,
-            warning=warning,
-        )
-
-    def _postprocess_prepare_case(
-        self,
-        case: object,
-        *,
-        raw_response: str,
-        needs_llm: bool,
-        primary_error: str,
-        warning: str,
-    ) -> tuple[object, str, str]:
-        """Parse the LLM response or apply fallback. Returns (case, warning, fallback_error)."""
-        if needs_llm and raw_response:
-            try:
-                case.generate_output = self._bindings.generate_output_cls(response=raw_response)
-                case = self._bindings.parse_result(case)
-                case = self._bindings.extract_main_html_single(case)
-            except Exception as exc:  # noqa: BLE001
-                primary_error = _append_warning(primary_error, str(exc))
-                logger.debug("Dripper parse/extract failed, applying {} fallback: {}", self.fallback, primary_error)
-                fallback_result = self._apply_fallback(case, primary_error)
-                warning = _append_warning(warning, fallback_result[1])
-                return fallback_result[0], warning, fallback_result[2]
-            return case, warning, ""
-        if needs_llm and not primary_error:
-            primary_error = "empty Dripper response"
-        fallback_result = self._apply_fallback(case, primary_error)
-        warning = _append_warning(warning, fallback_result[1])
-        return fallback_result[0], warning, fallback_result[2]
-
-    def _build_case(self, *, html: str, url: str | None, simplified_html: str, mapped_html: str) -> object:
-        case = self._bindings.case_cls(self._bindings.input_cls(raw_html=html, url=url))
-        if simplified_html or mapped_html:
-            case.process_data = self._bindings.process_data_cls(simpled_html=simplified_html, map_html=mapped_html)
-        return case
-
-    def _apply_fallback(self, case: object, primary_error: str) -> tuple[object, str, str]:
-        return _apply_fallback_extraction(self._bindings, self._fallback_handler, case, primary_error)
+# ---------------------------------------------------------------------------
+# HTML/case helper functions (promoted from DripperHTMLExtractionStage statics)
+# These are used by DripperHTMLLayoutTemplateStage and the split sub-modules.
+# ---------------------------------------------------------------------------
+
+
+def _sanitize_case_output_html(case: object) -> None:
+    """Strip XML-incompatible characters from the output main_html in place."""
+    output_data = getattr(case, "output_data", None)
+    if output_data is None:
+        return
+    main_html = getattr(output_data, "main_html", None)
+    if isinstance(main_html, str):
+        output_data.main_html = _strip_xml_incompatible_chars(main_html)
+
+
+def _get_processed_attr(case: object, attr: str) -> str:
+    """Return a string attribute from case.process_data, or ''."""
+    process_data = getattr(case, "process_data", None)
+    value = getattr(process_data, attr, "") if process_data is not None else ""
+    return value if isinstance(value, str) else ""
+
+
+def _case_has_item_ids(case: object) -> bool:
+    """Return True if the simplified or mapped HTML contains _item_id attributes."""
+    return "_item_id" in _get_processed_attr(case, "simpled_html") or "_item_id" in _get_processed_attr(
+        case, "map_html"
+    )
+
+
+def _count_item_ids(case: object) -> int:
+    """Return the number of distinct _item_id values in the simplified/mapped HTML."""
+    html = _get_processed_attr(case, "simpled_html") or _get_processed_attr(case, "map_html")
+    return len(set(_ITEM_ID_RE.findall(html)))
+
+
+def _coerce_html(value: object) -> str:
+    """Coerce an arbitrary HTML column value to a clean string."""
+    if _is_missing(value):
+        return ""
+    if isinstance(value, bytes | bytearray):
+        raw_bytes = bytes(value)
+        decoded = _decode_html_bytes(raw_bytes)
+        if decoded is None:
+            decoded = raw_bytes.decode("utf-8", errors="replace")
+        return _strip_xml_incompatible_chars(decoded or "")
+    return _strip_xml_incompatible_chars(str(value))
+
+
+def _coerce_optional_str(value: object) -> str | None:
+    """Coerce an arbitrary URL column value to a string or None."""
+    if _is_missing(value):
+        return None
+    text = str(value)
+    return text if text else None
+
+
+def _is_empty_document_error(error: str) -> bool:
+    """Return True if the error message indicates an empty/missing HTML document."""
+    normalized = error.lower()
+    return "document is empty" in normalized or "empty html tree" in normalized or "empty html input" in normalized
+
+
+def _generation_config_for_item_count(stage: Any, item_count: int) -> GenerationConfig:  # noqa: ANN401
+    """Compute a GenerationConfig scaled to item_count (shared by Extraction and Preprocess stages)."""
+    base = stage.generation_config or GenerationConfig()
+    if not stage.dynamic_max_tokens or base.max_tokens is None or item_count <= 0:
+        return base
+    dynamic_max_tokens = max(
+        stage.dynamic_min_max_tokens,
+        item_count * stage.dynamic_max_tokens_per_item + stage.dynamic_max_token_padding,
+    )
+    return replace(base, max_tokens=min(base.max_tokens, dynamic_max_tokens))
+
+
+# ---------------------------------------------------------------------------
+# DripperHTMLExtractionStage, DripperHTMLPreprocessStage,
+# DripperHTMLInferenceStage, DripperHTMLPostprocessStage
+# are defined in their own focused modules:
+#   extraction.py, preprocessing.py, inference.py
+# They are re-exported via __init__.py so external import paths are unchanged.
+# ---------------------------------------------------------------------------
 
 
 @dataclass(kw_only=True)
@@ -1958,7 +994,7 @@ class DripperHTMLLayoutTemplateStage(ProcessingStage[DocumentBatch, DocumentBatc
         for idx, row in df.iterrows():
             if not bool(row.get(_DRIPPER_NEEDS_LLM_COL, False)):
                 continue
-            html_text = DripperHTMLExtractionStage._coerce_html(row.get(self.html_col, ""))
+            html_text = _coerce_html(row.get(self.html_col, ""))
             if not html_text.strip():
                 continue
             try:
@@ -2017,7 +1053,7 @@ class DripperHTMLLayoutTemplateStage(ProcessingStage[DocumentBatch, DocumentBatc
         for idx, row in df.iterrows():
             if not bool(row.get(_DRIPPER_NEEDS_LLM_COL, False)):
                 continue
-            html_text = DripperHTMLExtractionStage._coerce_html(row.get(self.html_col, ""))
+            html_text = _coerce_html(row.get(self.html_col, ""))
             if not html_text.strip():
                 continue
             layout_key = self._row_layout_id_key(row)
@@ -2071,7 +1107,7 @@ class DripperHTMLLayoutTemplateStage(ProcessingStage[DocumentBatch, DocumentBatc
 
         samples: list[dict[str, Any]] = []
         for idx in indexes:
-            html_text = DripperHTMLExtractionStage._coerce_html(df.iloc[idx].get(self.html_col, ""))
+            html_text = _coerce_html(df.iloc[idx].get(self.html_col, ""))
             if not html_text.strip():
                 continue
             sample: dict[str, Any] = {"track_id": str(idx), "html": html_text}
@@ -2653,7 +1689,7 @@ class DripperHTMLLayoutTemplateStage(ProcessingStage[DocumentBatch, DocumentBatc
         candidates = [
             {
                 "track_id": str(idx),
-                "html": DripperHTMLExtractionStage._coerce_html(df.iloc[idx].get(self.html_col, "")),
+                "html": _coerce_html(df.iloc[idx].get(self.html_col, "")),
             }
             for idx in indexes
         ]
@@ -2683,7 +1719,7 @@ class DripperHTMLLayoutTemplateStage(ProcessingStage[DocumentBatch, DocumentBatc
         if inference_result.primary_error:
             return self._postprocess_error_row(row, inference_result, _InferContext(layout_cluster=cluster_id)), None
 
-        html_text = DripperHTMLExtractionStage._coerce_html(row.get(self.html_col, ""))
+        html_text = _coerce_html(row.get(self.html_col, ""))
         mapped_html = str(row.get(self.mapped_html_col, "") or "")
         case = self._build_case(row)
         try:
@@ -2754,7 +1790,7 @@ class DripperHTMLLayoutTemplateStage(ProcessingStage[DocumentBatch, DocumentBatc
         cluster_id: str,
     ) -> _LayoutTemplateRowResult:
         started = time.perf_counter()
-        html_text = DripperHTMLExtractionStage._coerce_html(row.get(self.html_col, ""))
+        html_text = _coerce_html(row.get(self.html_col, ""))
         mapped_html = str(row.get(self.mapped_html_col, "") or "")
         use_mapped_item_ids = (
             self.layout_template_propagation_target == "mapped_item_ids" and "_item_id" in mapped_html
@@ -3059,8 +2095,8 @@ class DripperHTMLLayoutTemplateStage(ProcessingStage[DocumentBatch, DocumentBatc
         )
 
     def _build_case(self, row: pd.Series) -> object:
-        html_text = DripperHTMLExtractionStage._coerce_html(row.get(self.html_col, ""))
-        url = DripperHTMLExtractionStage._coerce_optional_str(row.get(self.url_col) if self.url_col else None)
+        html_text = _coerce_html(row.get(self.html_col, ""))
+        url = _coerce_optional_str(row.get(self.url_col) if self.url_col else None)
         case = self._bindings.case_cls(self._bindings.input_cls(raw_html=html_text, url=url))
         simplified_html = str(row.get(self.simplified_html_col, "") or "")
         mapped_html = str(row.get(self.mapped_html_col, "") or "")
@@ -3071,10 +2107,7 @@ class DripperHTMLLayoutTemplateStage(ProcessingStage[DocumentBatch, DocumentBatc
     def _fallback_and_convert(self, row: pd.Series, *, primary_error: str = "") -> _DripperPostResult:
         started = time.perf_counter()
         case = self._build_case(row)
-        if (
-            bool(row.get(_DRIPPER_EMPTY_INPUT_COL, False))
-            or not DripperHTMLExtractionStage._coerce_html(row.get(self.html_col, "")).strip()
-        ):
+        if bool(row.get(_DRIPPER_EMPTY_INPUT_COL, False)) or not _coerce_html(row.get(self.html_col, "")).strip():
             return _DripperPostResult(
                 postprocess_time_s=time.perf_counter() - started,
                 warning=_append_warning(primary_error, "empty HTML input"),
@@ -3098,7 +2131,7 @@ class DripperHTMLLayoutTemplateStage(ProcessingStage[DocumentBatch, DocumentBatc
     def _convert_case(self, case: object, *, warning: str = "") -> _DripperPostResult:
         conversion_error = ""
         try:
-            DripperHTMLExtractionStage._sanitize_case_output_html(case)
+            _sanitize_case_output_html(case)
             case = self._bindings.convert2content(case, output_format=self.output_format)
         except Exception as exc:  # noqa: BLE001
             conversion_error = str(exc)
@@ -3111,7 +2144,7 @@ class DripperHTMLLayoutTemplateStage(ProcessingStage[DocumentBatch, DocumentBatc
             main_content = ""
         error = ""
         if conversion_error:
-            if DripperHTMLExtractionStage._is_empty_document_error(conversion_error) and not str(main_html).strip():
+            if _is_empty_document_error(conversion_error) and not str(main_html).strip():
                 warning = _append_warning(warning, conversion_error)
             else:
                 error = conversion_error
